data-center

data-center


See merge request !189

+++ smart-g60/device-monitor/src/main/java/com/pingok/devicemonitor/domain/smartToilet/TblSmartToiletCubicle.java
@@ -22,7 +22,7 @@ public class TblSmartToiletCubicle implements Serializable {
     /**
      * 厕所ID
      */
-    private Integer serId;
+    private Long serId;
 
     /**
      * 厕位编号
